% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_board.R
\name{define_board}
\alias{define_board}
\title{Define Board}
\usage{
define_board(
  height = 10,
  width = 10,
  ladder = list(c(1, 23), c(9, 36), c(50, 75), c(68, 98)),
  chute = list(c(26, 8), c(78, 48), c(35, 6), c(90, 69))
)
}
\arguments{
\item{height}{A positive integer representing the number of rows in the board}

\item{width}{A positive integer representing the number of columns in the board}

\item{ladder}{A list of vectors, each with two positive integers representing the start and end position of a ladder}

\item{chute}{A list of vectors, each with two positive integers representing the start and end position of a chute}
}
\value{
A list containing information of four user-defined inputs
}
\description{
This function returns a list containing information of a user-defined board for the chute-and-ladders game.
You can also:
Visualize the board with \code{show_board()} function.
Play a solo game with \code{play_solo()} function.
Perform simulation with \code{simualte_board()} function.
}
\examples{
board <- define_board(height=10, width=10,
ladder=list(c(1, 23), c(9,36), c(50,75), c(68,98)),
chute=list(c(26,8), c(78,48), c(35,6), c(90,69)))

}
