% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_solo.R
\name{play_solo}
\alias{play_solo}
\title{Play Solo}
\usage{
play_solo(theboard, verbose = FALSE)
}
\arguments{
\item{theboard}{A list of four elements containing board information:\code{height},\code{width},\code{ladder}and \code{chute}}

\item{verbose}{A logical T/F indicating whether to show the function output for each turn}
}
\value{
if verbose=F, return a list of play log: \code{turns}:the number of total turns, \code{ladder_tally}: the frequency of use of each ladder;
\code{chute_tally}:the frequency of use of each ladder; and \code{move_log}: record of all spaces that the user landed
if verbose=T, return a list of the above play log and displays each move of this play:
The start of each turn begins with the turn number: \verb{Turn x}
It announces where the player is located at the start of the turn: \verb{Start at x}
It announces what the player spun: \code{Spinner: x}
If the player lands on a special space, it says \verb{Landed on: x}
It will announce Chute! or Ladder! depending on what type of special space it is.
It announces where the player ends the turn: \verb{Turn ends at: x}
}
\description{
This function demonstrates a solo player chute-and-ladders game with a user defined board.
You can also:
Define the board with \code{define_board()} function.
Visualize the board with \code{show_board()} function.
Perform simulation with \code{simualte_board()} function.
}
\examples{
board <- define_board(height=10, width=10,
ladder=list(c(1, 23), c(9,36), c(50,75), c(68,98)),
chute=list(c(26,8), c(78,48), c(35,6), c(90,69)))

play_solo(board, verbose = TRUE)

}
