% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_board.R
\name{show_board}
\alias{show_board}
\title{Show Board}
\usage{
show_board(theboard)
}
\arguments{
\item{theboard}{A list of four elements of board information:\code{height},\code{width},\code{ladder}and \code{chute}}
}
\value{
This function does not return a value but produces a plot representing the game board.
}
\description{
This function visualizes the board of a chutes-and-ladders game with a user-defined board.
You can also:
Define the board with \code{define_board()} function.
Play a solo game with \code{play_solo()} function.
Perform simulation with \code{simualte_board()} function.
}
\examples{
board <- define_board(height=10, width=10,
ladder=list(c(1, 23), c(9,36), c(50,75), c(68,98)),
chute=list(c(26,8), c(78,48), c(35,6), c(90,69)))

show_board(board)

}
