% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_board.R
\name{simulate_board}
\alias{simulate_board}
\title{Boards Simulation}
\usage{
simulate_board(board, n = 10000)
}
\arguments{
\item{board}{A list created by define_board() containing game board information}

\item{n}{A positive large integer representing number of simulations}
}
\value{
A list of matrices containing simulation results of \code{AllTurns}, \code{AllLadders} and \code{AllChutes}
}
\description{
This function simulates a solo player Chutes-and-Ladders game.
For more information, you can:
Define the board with \code{define_board()} function.
Visualize the board with \code{show_board()} function.
Play a solo game with \code{play_solo()} function.
Perform simulation with \code{simualte_board()} function.
}
